requirements.txt
black==22.12.0
boto3==1.26.51
botocore==1.29.51
cachelib==0.9.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Login==0.6.2
flask-paginate==2022.1.8
Flask-Reuploaded==1.4.0
Flask-WTF==1.1.1
fonttools==4.39.2
greenlet==2.0.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mypy-extensions==0.4.3
numpy==1.24.2
packaging==23.0
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
# psycopg2==2.9.9
psycopg2-binary
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.17.0
